chore: apply all requested changes and cleanup
