+++ lib/presentation/bottom_navigation/views/bottom_navigation_view.dart
@@ -1,4 +1,5 @@
-import 'package:flowery_tracking_app/presentation/bottom_navigation/views/widgets/bottom_navigation_body.dart';
+import 'package:flowery_tracking_app/core/di/di.dart';
+import 'package:flowery_tracking_app/presentation/bottom_navigation/views/widgets/bottom_navigation_view_body.dart';
 import 'package:flowery_tracking_app/presentation/bottom_navigation/views_model/bottom_navigation_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -8,9 +9,9 @@ class BottomNavigationView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (_) => BottomNavigationCubit(),
-      child: const Scaffold(body: BottomNavigationBody()),
+    return BlocProvider<BottomNavigationCubit>(
+      create: (_) => getIt.get<BottomNavigationCubit>(),
+      child: const Scaffold(body: BottomNavigationViewBody()),
     );
   }
 }
